refactor: 쿠폰 컴포넌트화
